Move remote test to appropriate file
